+++ src/index.ts
@@ -1,11 +1,11 @@
 import style from './style.scss';
 import { main } from './main';
-import { version as supersplatVersion } from '../package.json';
+import { version as appVersion } from '../package.json';
 import { version as pcuiVersion, revision as pcuiRevision } from 'pcui';
 import { version as engineVersion, revision as engineRevision } from 'playcanvas';
 
 // print out versions of dependent packages
 // NOTE: add dummy style reference to prevent tree shaking
-console.log(`SuperSplat v${supersplatVersion} | PCUI v${pcuiVersion} (${pcuiRevision}) | PlayCanvas Engine v${engineVersion} (${engineRevision})${style.slice(0, 0)}`);
+console.log(`SuperSplat v${appVersion} | PCUI v${pcuiVersion} (${pcuiRevision}) | Engine v${engineVersion} (${engineRevision})${style.slice(0, 0)}`);
 
 main();
