+++ src/style.scss
@@ -37,6 +37,7 @@ body {
     flex-grow: 0;
     display: flex;
     flex-direction: column;
+    justify-content: space-between;
     border-right: 1px solid $bcg-darker;
 }
 
@@ -66,12 +67,6 @@ body {
     flex-shrink: 0;
     flex-grow: 0;
     margin: 5px;
-}
-
-#toolbar-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
 }
 
 .toolbar-button {
