+++ src/ui/toolbar.ts
@@ -1,24 +1,10 @@
-import { Button, Container, Label, Overlay, Panel } from 'pcui';
+import { Button, Container } from 'pcui';
 import logo from './playcanvas-logo.png';
-
-const shortcutList = [
-    { key: 'F', action: 'Focus camera' },
-    { key: 'I', action: 'Invert selection' },
-    { key: 'R', action: 'Toggle rect selection' },
-    { key: 'B', action: 'Toggle brush selection' },
-    { key: '[ ]', action: 'Decrease/Increase brush size' },
-    { key: 'Shift', action: 'Add to selection' },
-    { key: 'Ctrl', action: 'Remove from selection' },
-    { key: 'Delete', action: 'Delete selected splats' },
-    { key: 'Esc', action: 'Cancel rect selection' },
-    { key: 'Ctrl + Z', action: 'Undo' },
-    { key: 'Ctrl + Shift + Z', action: 'Redo' },
-    { key: 'Space', action: 'Toggle debug splat display' }
-];
+import { ShortcutsPopup } from './shortcuts';
 
 class Toolbar extends Container {
     constructor(appContainer: Container, args = {}) {
-        Object.assign(args, {
+        args = Object.assign(args, {
             id: 'toolbar-container'
         });
 
@@ -79,44 +65,9 @@ class Toolbar extends Container {
             window.open('https://github.com/playcanvas/super-splat', '_blank').focus();
         });
 
-        const shortcutsContainer = new Container({
-            id: 'shortcuts-container'
-        });
+        // keyboard shortcuts
+        const shortcutsPopup = new ShortcutsPopup();
 
-        shortcutList.forEach((shortcut) => {
-            const key = new Label({
-                class: 'shortcut-key',
-                text: shortcut.key
-            });
-
-            const action = new Label({
-                class: 'shortcut-action',
-                text: shortcut.action
-            });
-
-            const entry = new Container({
-                class: 'shortcut-entry'
-            });
-
-            entry.append(key);
-            entry.append(action);
-
-            shortcutsContainer.append(entry);
-        });
-
-        const shortcutsPanel = new Panel({
-            id: 'shortcuts-panel',
-            headerText: 'KEYBOARD SHORTCUTS'
-        });
-
-        shortcutsPanel.append(shortcutsContainer);
-
-        const shortcutsPopup = new Overlay({
-            id: 'shortcuts-popup',
-            clickable: true,
-            hidden: true
-        });
-        shortcutsPopup.append(shortcutsPanel);
         appContainer.append(shortcutsPopup);
 
         // keyboard shortcuts
@@ -131,12 +82,8 @@ class Toolbar extends Container {
         toolbarHelpContainer.append(shortcuts);
         toolbarHelpContainer.append(github);
 
-        // toolbar
-        const toolbarContainer = new Container({
-            id: 'toolbar-container'
-        });
-        toolbarContainer.append(toolbarToolsContainer);
-        toolbarContainer.append(toolbarHelpContainer);
+        this.append(toolbarToolsContainer);
+        this.append(toolbarHelpContainer);
     }
 }
